remove previous versions of Bouncy Castle
This will speed up the build, but its primary purpose is to make it
easier to change method signatures to work with the new KeyIdentifier
class introduced in 1.80.

We may need to introduce different source branches to allow us to work
with mutually incompatible versions of Bouncy Castle, but since we're
already making breaking changes in this branch, we may as well also drop
support for the older Bouncy Castle versions.